new file:   app/Http/Controllers/AlbumAdminController.php 	modified:   app/Http/Controllers/AlbumController.php 	modified:   app/Http/Controllers/SongController.php 	new file:   app/Http/Controllers/admin/SongAdminController.php 	modified:   app/Http/Controllers/admin/UserAdminController.php 	modified:   app/Models/Song.php 	modified:   routes/api.php 	modified:   vue/src/assets/base.css 	modified:   vue/src/components/AccountView/AccountViewSideBar.vue 	new file:   vue/src/components/UI/BaseTableBodyPagination.vue 	new file:   vue/src/components/icons/IconNextArrow.vue 	new file:   vue/src/components/icons/IconPreviousArrow.vue 	modified:   vue/src/model/songModel.ts 	modified:   vue/src/router/index.ts 	modified:   vue/src/store/module/adminModule.ts 	modified:   vue/src/views/AccountView/AccountAdmin/AccountAdminDashboard.vue 	new file:   vue/src/views/AccountView/AccountAdmin/AccountAdminDics/AccountAdminAlbums.vue 	new file:   vue/src/views/AccountView/AccountAdmin/AccountAdminDics/AccountAdminDiscsView.vue 	new file:   vue/src/views/AccountView/AccountAdmin/AccountAdminDics/AccountAdminSongEdit.vue 	new file:   vue/src/views/AccountView/AccountAdmin/AccountAdminDics/AccountAdminSongs.vue 	modified:   vue/src/views/AccountView/AccountAdmin/AccountAdminUsers/AccountAdminUserAlbum.vue 	modified:   vue/src/views/AccountView/AccountAdmin/AccountAdminUsers/AccountAdminUserEdit.vue 	modified:   vue/src/views/AccountView/AccountAdmin/AccountAdminView.vue 	modified:   vue/src/views/AccountView/AccountUpload/AccountUploadSong/AccountUploadSongEdit.vue

+++ app/Http/Controllers/admin/UserAdminController.php
@@ -5,12 +5,10 @@ namespace App\Http\Controllers\admin;
 use App\Models\Song;
 use App\Models\User;
 use App\Models\Album;
-use App\Models\Playlist;
 use App\Models\LikedSong;
 use App\Models\SongGenre;
 use App\Models\SongArtist;
 use App\Models\PlaylistSong;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
@@ -20,7 +18,7 @@ class UserAdminController extends Controller
     //
     public function getAll()
     {
-        if (Auth::user()->role !== 'Admin') {
+        if (Auth::user()->role === '') {
             return response()->json(['message' => 'You are not authorized to access this resource.'], 403);
         } else {
             $users = User::get();
@@ -33,7 +31,7 @@ class UserAdminController extends Controller
 
     public function show($id)
     {
-        if (Auth::user()->role !== 'Admin') {
+        if (Auth::user()->role === '') {
             return response()->json(['message' => 'You are not authorized to access this resource.'], 403);
         } else {
             $user = User::find($id);
@@ -53,6 +51,15 @@ class UserAdminController extends Controller
 
     public function showUserUploadSong($id)
     {
+        if (Auth::user()->role === '') {
+            return response()->json(
+                [
+                    'status' => 'error',
+                    'message' => 'You are not authorized to access this resource.'
+                ],
+                403
+            );
+        }
         $songs = DB::select(
             'SELECT songs.*,
                     artists.name AS artist_name,artists.artist_id AS artist_id,
@@ -71,9 +78,35 @@ class UserAdminController extends Controller
         ]);
     }
 
+    public function updateRole($id)
+    {
+        if (Auth::user()->role !== 'Admin') {
+            return response()->json(['message' => 'You are not authorized to access this resource.'], 403);
+        } else {
+            $user = User::find($id);
+            if (!$user) {
+                return response()->json([
+                    'message' => 'User not found',
+                    'status' => 'error'
+                ], 404);
+            } else {
+                if ($user->role === 'Mod') {
+                    $user->role = '';
+                } else {
+                    $user->role = 'Mod';
+                }
+                $user->save();
+                return response()->json([
+                    'message' => 'Update role successfully',
+                    'status' => 'success'
+                ], 200);
+            }
+        }
+    }
+
     public function deleteUser($id)
     {
-        if (Auth::user()->role !== 'Admin' && Auth::user()->user_id === $id) {
+        if (Auth::user()->role === '' && Auth::user()->user_id === $id) {
             return response()->json(
                 [
                     'status' => 'error',
@@ -104,6 +137,12 @@ class UserAdminController extends Controller
                 //     $playlist->song()->detach();
                 // }
                 // Playlist::where('user_id', $id)->delete();
+                if ($user->role === 'Admin') {
+                    return response()->json([
+                        'message' => 'You can not delete admin',
+                        'status' => 'error'
+                    ], 403);
+                }
                 DB::table('personal_access_tokens')->where('tokenable_id', $id)->delete();
                 $user->delete();
                 return response()->json([
@@ -116,11 +155,21 @@ class UserAdminController extends Controller
 
     public function showUserUploadAlbum($id)
     {
+        if (Auth::user()->role === '') {
+            return response()->json(
+                [
+                    'status' => 'error',
+                    'message' => 'You are not authorized to access this resource.'
+                ],
+                403
+            );
+        }
         $albumUploaded = DB::table('albums')
             ->leftJoin('songs', 'albums.album_id', '=', 'songs.album_id')
             ->select('albums.*', DB::raw('count(songs.song_id) as songCount'))
             ->where('albums.user_id', Auth::user()->user_id)
             ->groupBy('albums.id')
+            ->where('albums.user_id', $id)
             ->get();
         return response()->json([
             'status' => 'success',
@@ -128,25 +177,36 @@ class UserAdminController extends Controller
         ]);
     }
 
-    public function deleteSong($id)
+
+
+    public function deleteAlbum($id)
     {
-        $song = Song::find($id);
-        if (!$song) {
-            return response()->json([
-                'message' => 'Song not found',
-                'status' => 'error'
-            ], 404);
-        } else {
-            SongArtist::where('song_id', $id)->delete();
-            SongGenre::where('song_id', $id)->delete();
-            PlaylistSong::where('song_id', $id)->delete();
-            LikedSong::where('song_id', $id)->delete();
-            @unlink(public_path('storage/upload/song_src/') . $id . '.ogg');
-            $song->delete();
-            return response()->json([
-                'message' => 'Song deleted successfully',
-                'status' => 'success'
-            ], 200);
+        if (Auth::user()->role === '') {
+            return response()->json(
+                [
+                    'status' => 'error',
+                    'message' => 'You are not authorized to access this resource.'
+                ],
+                403
+            );
         }
+        $album = Album::find($id);
+        if (!$album->id) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Album not found'
+            ]);
+        }
+        @unlink(public_path('storage/upload/album_cover/') . "/" . $album->album_id);
+        $song = Song::where('album_id', $album->album_id)->get();
+        foreach ($song as $s) {
+            $s->album_id = null;
+            $s->save();
+        }
+        $album->delete();
+        return response()->json([
+            'status' => 'success',
+            'message' => 'Album deleted successfully!'
+        ]);
     }
 }
